pyproject.toml
[project]
name = "prop-firm-dragon-funded"
version = "0.1.0"
description = "Customer support RAG bot for the Dragon Funded forex prop firm."
readme = "README.md"
authors = [{ name = "Dragon Funded AI Ops" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "chromadb>=0.5.3",
    "fastapi>=0.110.0",
    "google-generativeai>=0.5.0",
    "langchain-community>=0.2.0",
    "langchain-google-genai>=0.0.6",
    "langchain>=0.2.0",
    "langgraph>=0.1.6",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.2.0",
    "uvicorn[standard]>=0.29.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]

[tool.uv]
package = true

[tool.setuptools]
packages = { find = { exclude = ["storage*", "tests*"] } }

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
